refactor markdown relative components
