--- components/Markdown/Markdown.tsx
@@ -1,3 +1,0 @@
-export { default as MarkdownRender } from './MarkdownRender';
-export { default as TuiEditor } from './TuiEditor';
-export { default as TuiEditorWrapper } from './TuiEditorWrapper';
